Rename factory module
